switch from setuptools to uv_build for sheep
