+++ software/python-package/shepherd_sheep/shared_mem_iv_output.py
@@ -9,7 +9,7 @@ from shepherd_core.data_models import PowerTracing
 from typing_extensions import Self
 
 from . import commons
-from . import sysfs_interface as sfs
+from . import sysfs_interface as sysfs
 from .logger import log
 from .shared_mem_iv_input import IVTrace
 
@@ -32,9 +32,9 @@ class SharedMemIVOutput:
 
     def __init__(self, mem_map: mmap, cfg: PowerTracing | None, ts_xp_start_ns: int) -> None:
         self._mm: mmap = mem_map
-        self.size_by_sys: int = sfs.get_trace_iv_out_size()
-        self.address: int = sfs.get_trace_iv_out_address()
-        self.base: int = sfs.get_trace_iv_inp_address()
+        self.size_by_sys: int = sysfs.get_trace_iv_out_size()
+        self.address: int = sysfs.get_trace_iv_out_address()
+        self.base: int = sysfs.get_trace_iv_inp_address()
 
         if self.size_by_sys != self.SIZE_SECTION:
             msg = f"[{type(self).__name__}] Size does not match PRU-data"
